+++ Program/data/objects/ModelScale.cs
@@ -2,11 +2,11 @@
 
 namespace Rejestracja.Data.Objects {
     class ModelScale {
-        public long id;
+        public int id;
         public String name;
         public int displayOrder;
 
-        public ModelScale(long id, String name, int displayOrder) {
+        public ModelScale(int id, String name, int displayOrder) {
             this.id = id;
             this.name = name;
             this.displayOrder = displayOrder;
